Removed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@3483 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/state_manager.cpp
@@ -286,12 +286,12 @@ namespace StateManager
                 SpinnerWidget* gauge = getCurrentScreen()->getWidget<SpinnerWidget>("sfx_volume");
                 assert(gauge != NULL);
 
-                gauge->setValue( sfx_manager->getMasterSFXVolume()*10.0f );
+                gauge->setValue( (int)(sfx_manager->getMasterSFXVolume()*10.0f) );
 
 
                 gauge = getCurrentScreen()->getWidget<SpinnerWidget>("music_volume");
                 assert(gauge != NULL);
-                gauge->setValue( sound_manager->getMasterMusicVolume()*10.f );
+                gauge->setValue( (int)(sound_manager->getMasterMusicVolume()*10.f) );
 
                 // ---- music volume
                 CheckBoxWidget* sfx = getCurrentScreen()->getWidget<CheckBoxWidget>("sfx_enabled");
